+++ megalinter/tests/test_megalinter/mega_linter_3_sarif_test.py
@@ -12,7 +12,7 @@ import uuid
 from megalinter import Linter, MegaLinter
 from megalinter.constants import DEFAULT_SARIF_REPORT_FILE_NAME
 from megalinter.reporters.SarifReporter import SarifReporter
-from megalinter.tests.test_megalinter.helpers import utilstest
+from megalinter import utilstest
 
 root = (
     os.path.dirname(os.path.abspath(__file__))